make MBItem._determine_name() an abstract method to children can better handle their naming
